requirements.txt
requests
pyyaml
click
